Move midnight_UTC from SM2 to base Scheduler
This function is not specific to the SM2 scheduler implementation,
it's more like a utility function. It's also called directly by
several scheduler consumers.

Move it to the base scheduler class to clarify things
